// src/api/generated/models/ApiResponsePageLeadResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { PageLeadResponse } from './PageLeadResponse';
export type ApiResponsePageLeadResponse = {
    success?: boolean;
    message?: string;
    data?: PageLeadResponse;
};